+++ R/performance_lrt.R
@@ -1,9 +1,13 @@
 #' @title Likelihood-Ratio-Test for Model Comparison
 #' @name performance_lrt
 #'
-#' @description Compute Likelihood-Ratio-Test (LRT) for model comparison. For regression models, this is obtained through the \code{anova(..., test="LRT")} function. This test gives usually somewhat close results (if not equivalent) to the commonly used F-test (obtained by default with \code{anova(...)}). The LRT is a maximum likelihood test used to see which model is a better (more likely) explanation of the data. Maximum likelihood tests make stronger assumptions than method of moments tests like the F-test, and in return are more efficient.
+#' @description Compute Likelihood-Ratio-Test (LRT) for model comparison, which tests which model is a better (more likely) explanation of the data. Likelihood-Ratio-Test (LRT) gives usually somewhat close results (if not equivalent) to the commonly used F-test (obtained by default with \code{anova(...)}). Maximum likelihood tests make stronger assumptions than method of moments tests like the F-test, and in turn are more efficient.
+#' \itemize{
+#' \item For regression models, this is similar to \code{anova(..., test="LRT")} or \code{lmtest::lrtest(...)}, with one major difference: all models are tested against the same reference model (the first one), instead of testing each model to its previous neighbour.
+#' \item For \code{lavaan} models (SEM, CFA), the function calls \code{lavaan::lavTestLRT()}.
+#' }
 #'
-#' @param ... Multiple model objects, which should respond to \code{anova()}.
+#' @param ... Multiple model objects.
 #'
 #' @return A data frame, based on the results from \code{anova()}.
 #'
@@ -12,11 +16,13 @@
 #' @seealso \code{\link[=compare_performance]{compare_performance()}} to compare performance of many different models.
 #'
 #' @examples
+#' # Regression Models
 #' m1 <- lm(mpg ~ wt + cyl, data = mtcars)
 #' m2 <- lm(mpg ~ wt + cyl + gear, data = mtcars)
 #' m3 <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
 #' performance_lrt(m1, m2, m3)
 #'
+#' # Lavaan Models
 #' if (require("lavaan")) {
 #'   structure <- " visual  =~ x1 + x2 + x3
 #'                  textual =~ x4 + x5 + x6
@@ -36,10 +42,12 @@ performance_lrt <- function(...) {
 }
 
 
+
+#' @param estimator Applied when comparing regression models. Corresponds to the different estimators for the standard deviation of the errors. If \code{estimator="OLS"} (default), then it uses the same method as \code{anova(..., test="LRT")} implemented in base R, i.e., scaling by n−k (the unbiased OLS estimator) and using this estimator under the alternative hypothesis. If \code{estimator="ML"}, which is for instance used by \code{lrtest(...)} in the \code{lmtest}, the scaling is done by n (the biased ML estimator) and the estimator under the null hypothesis. In moderately large samples, the differences should be negligible, but it is possible that OLS would perform slightly better in small samples with Gaussian errors.
 #' @importFrom stats anova
 #' @importFrom insight is_model
 #' @export
-performance_lrt.default <- function(...) {
+performance_lrt.default <- function(..., estimator="OLS") {
   objects <- list(...)
 
   if (!all(sapply(objects, insight::is_model))) {
@@ -47,25 +55,17 @@ performance_lrt.default <- function(...) {
   }
 
   object_names <- match.call(expand.dots = FALSE)$`...`
+  names(objects) <- object_names
 
   # LRT for model comparison
   if (length(objects) > 1) {
-    # sd_mle <- function(model) sqrt(mean(residuals(model)^2))
-    # ll <- function(model, sd) {
-    #   sum(dnorm(insight::get_response(model), mean = fitted(model), sd = sd, log = TRUE))
-    # }
-    # -2 * (ll(m2, sd_mle(m2)) - ll(m3, sd_mle(m3)))
-    lrt <- stats::anova(..., test = "LRT")
-
-    # sort
-    lrt_order <- order(order(rownames(lrt)))
-    objects <- objects[lrt_order]
-    object_names <- object_names[lrt_order]
-
-    .performance_lrt(objects, object_names, lrt)
+    out <- .performance_lrt_lm(objects, estimator=estimator)
   } else {
     warning("At least two models required for a Likelihood-Ratio-Test.", call. = FALSE)
   }
+
+  class(out) <- c("performance_lrt", "see_performance_lrt", "data.frame")
+  out
 }
 
 
@@ -105,12 +105,12 @@ performance_lrt.lavaan <- function(...) {
 
 .performance_lrt <- function(objects, object_names, lrt) {
   # create data frame with info about model name and class
-  m <- mapply(function(.x, .y) {
+  names_types <- do.call(rbind, mapply(function(.x, .y) {
     data.frame(Model = as.character(.y), Type = class(.x)[1], stringsAsFactors = FALSE)
-  }, objects, object_names, SIMPLIFY = FALSE)
+  }, objects, object_names, SIMPLIFY = FALSE))
 
   # bind all data
-  out <- cbind(do.call(rbind, m), lrt)
+  out <- cbind(names_types, lrt)
 
   # Drop cols
   out <- out[!names(out) %in% c("AIC", "BIC", "Df diff", "Chisq diff")]  # For regressions
@@ -169,9 +169,18 @@ performance_lrt.lavaan <- function(...) {
   out <- .performance_lrt_lm_pairwise(ref, ref, estimator = estimator)
   out$p <- out$Chi2 <- NA
 
-  out <- rbind(out, t(sapply(objects[-ref_model], .performance_lrt_lm_pairwise, ref = ref, estimator = estimator)))
+  for(model in objects[-ref_model]){
+    out <- rbind(out, .performance_lrt_lm_pairwise(model, ref = ref, estimator = estimator))
+  }
+
+  # Get model Type
+  names_types <- do.call(rbind, mapply(function(.x, .y) {
+    data.frame(Model = as.character(.y), Type = class(.x)[1], stringsAsFactors = FALSE)
+  }, objects, names(objects), SIMPLIFY = FALSE))
+
+  # Add Names
+  out <- cbind(names_types, out)
   row.names(out) <- NULL
-  out <- cbind(data.frame(Model = names(objects), stringsAsFactors = FALSE), out)
 
   out
 }
@@ -193,17 +202,18 @@ performance_lrt.lavaan <- function(...) {
   }
 
   # Get Statistic
+  # TODO: Could use a possible insight::get_residuals()
   if (tolower(estimator) == "ols") {
     estim <- function(object) sqrt(sum(stats::residuals(object)^2) / stats::df.residual(object))
     loglik <- loglik_function(ref, estim(model))
     stat <- -2 * (loglik - loglik_function(model, estim(model)))
   } else{
     estim <- function(object) sqrt(mean(stats::residuals(object)^2))
-    loglik <- loglik_function(ref, estim(ref))
-    stat <- -2 * (loglik - loglik_function(model, estim(model)))
+    stat <- -2 * (loglik_function(ref, estim(ref)) - loglik_function(model, estim(model)))
+    loglik <- as.numeric(stats::logLik(model)) # This is the loglik returned
   }
   # Get p-value
-  p <- stats::pchisq(stat, abs(df_diff), lower.tail = FALSE)
+  p <- stats::pchisq(abs(stat), abs(df_diff), lower.tail = FALSE)
 
   # Out
   data.frame(