+++ tests/testthat/test-compare_performance.R
@@ -8,18 +8,18 @@ if (require("testthat") && require("performance")) {
   test_that("compare_performance", {
     expect_equal(
       colnames(compare_performance(lm1, lm2, lm3)),
-      c("Model", "Type", "AIC", "BIC", "BF", "R2", "R2_adjusted", "RMSE", "Sigma", "df", "p")
+      c("Model", "Type", "AIC", "BIC", "BF", "R2", "R2_adjusted", "RMSE", "Sigma", "Chi2", "df", "p")
     )
 
     expect_warning(
       expect_equal(
         colnames(compare_performance(lm1, lm2, lm3, lm4)),
-        c("Model", "Type", "AIC", "BIC", "R2", "R2_adjusted", "RMSE", "Sigma")
+        c("Model", "Type", "AIC", "BIC", "R2", "R2_adjusted", "RMSE", "Sigma", "Chi2", "df", "p")
       )
     )
     expect_equal(
       colnames(compare_performance(lm1, lm2, lm3, lm4, verbose = FALSE)),
-      c("Model", "Type", "AIC", "BIC", "R2", "R2_adjusted", "RMSE", "Sigma")
+      c("Model", "Type", "AIC", "BIC", "R2", "R2_adjusted", "RMSE", "Sigma", "Chi2", "df", "p")
     )
   })
 }